+++ src/store/actionTypes.ts
@@ -12,3 +12,4 @@ export const SET_STATUS = 'SET_STATUS';
 export const SET_COMPLETE_STATUS = 'SET_COMPLETE_STATUS';
 export const SET_LOADING = 'SET_LOADING';
 export const SET_VALIDITY = 'SET_VALIDITY';
+export const SET_NEW_POOL_SUPPLY = 'SET_NEW_POOL_SUPPLY';
